bruk fail i arbeidsforhold og virksomhetløser
