+++ felles/src/main/kotlin/no/nav/helsearbeidsgiver/felles/Fail.kt
@@ -1,0 +1,58 @@
+package no.nav.helsearbeidsgiver.felles
+
+import com.fasterxml.jackson.annotation.JsonIgnore
+import com.fasterxml.jackson.databind.JsonNode
+import no.nav.helse.rapids_rivers.JsonMessage
+import no.nav.helse.rapids_rivers.MessageContext
+import no.nav.helse.rapids_rivers.River
+import no.nav.helsearbeidsgiver.felles.json.customObjectMapper
+
+data class Fail(
+    @JsonIgnore
+    val eventName: EventName?,
+    val behov: BehovType?,
+    val feilmelding: String,
+    val data: HashMap<DataFelt, Any>?,
+    val uuid: String?,
+    val forespørselId: String?
+) {
+    fun toJsonMessage(): JsonMessage {
+        return JsonMessage.newMessage(
+            mutableMapOf<String, Any>().putIfNotEmpty(
+                Key.EVENT_NAME.str to this.eventName,
+                Key.FAIL.str to this,
+                Key.UUID.str to this.uuid
+            )
+        )
+    }
+}
+
+@Suppress("UNCHECKED_CAST")
+fun <K, V> MutableMap<K, V>.putIfNotEmpty(vararg pair: Pair<K, V?>): MutableMap<K, V> {
+    this.putAll(pair.filter { it.second != null } as Iterable<Pair<K, V>>)
+    return this
+}
+
+fun JsonMessage.toFeilMessage(): Fail {
+    return customObjectMapper().treeToValue(this[Key.FAIL.str], Fail::class.java).copy(eventName = EventName.valueOf(this[Key.EVENT_NAME.str].asText()))
+}
+
+fun JsonMessage.createFail(feilmelding: String, data: HashMap<DataFelt, Any>? = null, behoveType: BehovType? = null): Fail {
+    val behovNode: JsonNode? = this.valueNullable(Key.BEHOV)
+    // behovtype trenger å vare definert eksplisit da behov elemente er en List
+    val behov: BehovType? = behoveType ?: if (behovNode != null) BehovType.valueOf(behovNode.asText()) else null
+    val forespørselId = this.valueNullableOrUndefined(Key.FORESPOERSEL_ID)?.asText()
+    val eventName = this.valueNullableOrUndefined(Key.EVENT_NAME)?.asText()
+    return Fail(eventName?.let { EventName.valueOf(eventName) }, behov, feilmelding, data, this.valueNullable(Key.UUID)?.asText(), forespørselId)
+}
+
+fun River.PacketListener.publishFail(fail: Fail, jsonMessage: JsonMessage, context: MessageContext) {
+    val message = JsonMessage.newMessage(
+        mapOf(
+            Key.EVENT_NAME.str to jsonMessage[Key.EVENT_NAME.str].asText(),
+            Key.FAIL.str to customObjectMapper().writeValueAsString(fail),
+            Key.UUID.str to jsonMessage[Key.UUID.str].asText()
+        )
+    )
+    context.publish(fail.toJsonMessage().toJson())
+}
