+++ pdl/src/main/kotlin/no/nav/helsearbeidsgiver/inntektsmelding/pdl/FulltNavnLøser.kt
@@ -5,18 +5,16 @@ package no.nav.helsearbeidsgiver.inntektsmelding.pdl
 import kotlinx.coroutines.runBlocking
 import no.nav.helse.rapids_rivers.JsonMessage
 import no.nav.helse.rapids_rivers.MessageContext
-import no.nav.helse.rapids_rivers.MessageProblems
 import no.nav.helse.rapids_rivers.RapidsConnection
 import no.nav.helse.rapids_rivers.River
 import no.nav.helsearbeidsgiver.felles.BehovType
 import no.nav.helsearbeidsgiver.felles.DataFelt
-import no.nav.helsearbeidsgiver.felles.Feil
 import no.nav.helsearbeidsgiver.felles.Feilmelding
 import no.nav.helsearbeidsgiver.felles.Key
 import no.nav.helsearbeidsgiver.felles.NavnLøsning
 import no.nav.helsearbeidsgiver.felles.PersonDato
 import no.nav.helsearbeidsgiver.felles.createFail
-import no.nav.helsearbeidsgiver.felles.json.customObjectMapper
+import no.nav.helsearbeidsgiver.felles.publishFail
 import no.nav.helsearbeidsgiver.pdl.PdlClient
 import org.slf4j.LoggerFactory
 import java.time.LocalDate
@@ -59,7 +57,7 @@ class FulltNavnLøser(
             logger.error("Klarte ikke hente navn for id $id")
             sikkerlogg.error("Det oppstod en feil ved henting av identitetsnummer: $identitetsnummer: ${ex.message} for id: $id", ex)
             publish(NavnLøsning(error = Feilmelding("Klarte ikke hente navn")), packet, context)
-            publishFail(packet.createFail("Ukjent person", behoveType = BehovType.FULLT_NAVN), context)
+            publishFail(packet.createFail("Ukjent person", behoveType = BehovType.FULLT_NAVN), packet, context)
         }
     }
 
@@ -82,21 +80,6 @@ class FulltNavnLøser(
         context.publish(message.toJson())
     }
 
-    fun publishFail(feil: Feil, context: MessageContext) {
-        context.publish(feil.toJsonMessage().toJson())
-    }
-
-    fun publishFail(fail: Feilmelding, jsonMessage: JsonMessage, context: MessageContext) {
-        val message = JsonMessage.newMessage(
-            mapOf(
-                Key.EVENT_NAME.str to jsonMessage[Key.EVENT_NAME.str].asText(),
-                Key.FAIL.str to customObjectMapper().writeValueAsString(fail),
-                Key.UUID.str to jsonMessage[Key.UUID.str].asText()
-            )
-        )
-        context.publish(message.toJson())
-    }
-
     suspend fun hentPersonInfo(identitetsnummer: String): PersonDato {
         val liste = pdlClient.fullPerson(identitetsnummer)?.hentPerson
         val fødselsdato: LocalDate? = liste?.foedsel?.firstOrNull()?.foedselsdato
@@ -109,6 +92,4 @@ class FulltNavnLøser(
             nøkkel.name to data
         )
     }
-
-    override fun onError(problems: MessageProblems, context: MessageContext) {}
 }
